meta-resin-digi-connectcore: u-boot-digi: Modify for BalenaOS
This commit adds the following changes:

* Add BalenaOS partition table

The above is quick and dirty. In future the partitioning scheme will be
changed to use fewer partitions and more volumes.

* Allow to set UBI max BEB

UBI calculates the amount of blocks it needs for bad block handling
out of the total size of the chip, and then allocated those per
partition. The default is 20 per 1024 bytes.

This commit allows to specify the max BEB size on the u-boot
environment so we can use smaller partitions.

This wouldn't be needed if we used bigger partitions and more volumes.

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>